refactor(#2449): extract shared map-canvas primitives; buildings + battles consume
frontend/src/map-canvas/: MapCanvasShell (ReactFlowProvider + ReactFlow +
Background/Controls + empty-state slot), useMapNodeInteraction (shared
role/tabIndex/keydown/click wiring for RoomNode/PlaceNode), coords.ts (grid
CELL/cellToPosition/positionToCell + plane computeBounds/scaleFor/
planeToCanvas/radiusToPixels/centeredNodePosition, one documented y-negation
convention), edges.ts (exitEdges). Moved gridMath.test.ts/mapMath.test.ts
coverage for the moved functions to map-canvas; buildings' gridMath.ts keeps
DIRECTIONS/PlacedRoom/GhostCell/ghostCells (dig-specific). BuilderCanvas and
BattleMapCanvas now consume MapCanvasShell — behavior-preserving refactor,
collapsed each canvas's Outer/Inner split since the shell now owns
ReactFlowProvider.
