+++ frontend/src/buildings/components/RoomNode.tsx
@@ -12,6 +12,7 @@ import { Handle, Position } from '@xyflow/react';
 import type { Node, NodeProps } from '@xyflow/react';
 
 import { Badge } from '@/components/ui/badge';
+import { useMapNodeInteraction } from '@/map-canvas/useMapNodeInteraction';
 
 import type { GhostCell } from '../gridMath';
 import type { ManagerRoom } from '../types';
@@ -26,20 +27,13 @@ export type RoomNodeType = Node<RoomNodeData>;
 
 function RoomNodeComponent({ data }: NodeProps<RoomNodeType>) {
   const { room, selected } = data;
+  const interaction = useMapNodeInteraction({ onSelect: () => data.onSelect(room.id) });
   return (
     <div
-      role="button"
-      tabIndex={0}
+      {...interaction}
       className={`h-[104px] w-[104px] cursor-pointer overflow-hidden rounded-md border bg-card p-2 shadow-sm transition-colors hover:border-primary/60 ${
         selected ? 'border-primary ring-2 ring-primary/40' : 'border-border'
       }`}
-      onClick={() => data.onSelect(room.id)}
-      onKeyDown={(event) => {
-        if (event.key === 'Enter' || event.key === ' ') {
-          event.preventDefault();
-          data.onSelect(room.id);
-        }
-      }}
       data-testid="builder-room-node"
       data-room-id={room.id}
     >
