+++ frontend/src/buildings/gridMath.ts
@@ -1,13 +1,17 @@
 /**
- * Pure grid math for the building builder canvas (#670 PR2).
+ * Pure grid math specific to the building builder canvas (#670 PR2) — dig
+ * directions and ghost-cell (empty adjacent cell) computation. The shared
+ * grid coordinate conversions (`CELL`, `cellToPosition`/`positionToCell`)
+ * and exit-edge pairing (`exitEdges`) moved to `@/map-canvas` (#2449) so the
+ * world-builder canvas can reuse them too.
  *
  * The backend stores building-local cells (grid_x, grid_y, floor) where
- * north is +grid_y. Screens draw y downward, so rendering negates grid_y.
+ * north is +grid_y — see `@/map-canvas/coords` for the y-negation convention.
  * One cell per room regardless of size — size renders as a label, never
  * as footprint.
  */
 
-export const CELL = 120;
+import type { Cell } from '@/map-canvas/coords';
 
 /** Planar dig directions, mirroring world/buildings/room_constants.py. */
 export const DIRECTIONS: Record<string, { dx: number; dy: number }> = {
@@ -21,11 +25,6 @@ export const DIRECTIONS: Record<string, { dx: number; dy: number }> = {
   southwest: { dx: -1, dy: -1 },
 };
 
-export interface Cell {
-  x: number;
-  y: number;
-}
-
 export interface PlacedRoom {
   id: number;
   grid_x: number | null;
@@ -36,34 +35,6 @@ export interface PlacedRoom {
 export interface GhostCell extends Cell {
   fromRoomId: number;
   direction: string;
-}
-
-export interface ExitRecord {
-  id: number;
-  name: string;
-  from_room_id: number;
-  to_room_id: number;
-}
-
-export interface ExitEdge {
-  /** Stable per room pair: `exit-<lowRoomId>-<highRoomId>`. */
-  id: string;
-  source: number;
-  target: number;
-  /** Exit leaving `source` toward `target` (if any). */
-  there: ExitRecord | null;
-  /** Exit leaving `target` back toward `source` (if any). */
-  back: ExitRecord | null;
-}
-
-/** Grid cell -> React Flow canvas position (top-left of the node). */
-export function cellToPosition(cell: Cell): { x: number; y: number } {
-  return { x: cell.x * CELL, y: -cell.y * CELL };
-}
-
-/** Canvas position -> nearest grid cell (inverse of cellToPosition). */
-export function positionToCell(position: { x: number; y: number }): Cell {
-  return { x: Math.round(position.x / CELL), y: Math.round(-position.y / CELL) };
 }
 
 const cellKey = (x: number, y: number) => `${x},${y}`;
@@ -89,30 +60,4 @@ export function ghostCells(rooms: PlacedRoom[], floor: number): GhostCell[] {
     }
   }
   return [...ghosts.values()];
-}
-
-/**
- * Pair directed exits into one edge per room pair. A one-way exit still
- * yields an edge (its `back` stays null).
- */
-export function exitEdges(exits: ExitRecord[]): ExitEdge[] {
-  const edges = new Map<string, ExitEdge>();
-  for (const exit of exits) {
-    const [low, high] = [
-      Math.min(exit.from_room_id, exit.to_room_id),
-      Math.max(exit.from_room_id, exit.to_room_id),
-    ];
-    const id = `exit-${low}-${high}`;
-    let edge = edges.get(id);
-    if (!edge) {
-      edge = { id, source: exit.from_room_id, target: exit.to_room_id, there: null, back: null };
-      edges.set(id, edge);
-    }
-    if (exit.from_room_id === edge.source) {
-      edge.there = edge.there ?? exit;
-    } else {
-      edge.back = edge.back ?? exit;
-    }
-  }
-  return [...edges.values()];
 }
